requirements.txt
Flask==3.1.2
python-dotenv==1.2.1
requests==2.32.5
networkx==3.4.2
pandas==2.3.3
reportlab==4.4.7
google-genai==1.56.0
matplotlib==3.10.8
Pillow==12.0.0
plotly==6.5.0
kaleido==1.2.0
SQLAlchemy==2.0.45
psycopg2-binary==2.9.11
celery==5.6.0
redis==7.1.0
scikit-learn==1.7.2
numpy==2.2.6
xgboost==3.1.2
web3==6.11.0
aiohttp==3.9.1
aioredis==2.0.1
python-socketio==5.10.0
python-engineio==4.8.0
APScheduler==3.10.4
graphql-core==3.2.3
# slither-analyzer==0.9.3